minor season improvement and refactor
